daily/requirements.txt
asgiref
Babel
beautifulsoup4
colorama
Django
django-bootstrap-v5
django-cors-headers
django-filter
djangorestframework
iniconfig
Markdown
packaging
pluggy
pytest
pytz
soupsieve
sqlparse
tzdata
django-babel
requests

